+++ src/Units/CaravanActions.cs
@@ -83,7 +83,9 @@ namespace CivOne.Units
 					$"arrives in {city.Name}",
 					"Trade route established",
 					$"Revenue: ${revenue}."));
-			Game.Instance.GetPlayer(unit.Owner).Gold += (short)revenue;
+			// No (short) cast. Revenue is unbounded — the reported case was 10,000 — and the
+			// cast is what truncated the sum before Player.Gold's clamp could see it.
+			Game.Instance.GetPlayer(unit.Owner).Gold += revenue;
 			Game.Instance.DisbandUnit(unit);
 		}
